Hackathone/backend/ui_chatbot/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS support
python-dotenv==1.0.0

# Existing backend dependencies (imported)
# qdrant-client, cohere, openai are in parent directory
